chore: make it better🪴

+++ src/draw_uml_backend/server_automation/server_schema_generation.py
@@ -1,3 +1,4 @@
+from pathlib import Path
 import os
 
 
@@ -12,7 +13,8 @@ def write_to_output2(content: str):
 
 
 def create_server_setup():
-    write_to_output1("""
+    write_to_output1(
+        """
 from fastapi.middleware.cors import CORSMiddleware
 from typing import List
 import schema
@@ -46,14 +48,17 @@ app.add_middleware(
 async def read_root():
     response = RedirectResponse(url="/docs")
     return response  
-""")
+"""
+    )
 
 
 def create_db_schema_setup():
-    write_to_output2("""
+    write_to_output2(
+        """
 from typing import Union, Optional, List
 from pydantic import BaseModel
-""")
+"""
+    )
 
 
 def convert_to_http_method(method: str):
@@ -84,7 +89,8 @@ def create_server_methods(method: str, resource: str):
     res = "{"
     source = "}"
     if method == "create":
-        write_to_output1(f"""
+        write_to_output1(
+            f"""
 
 # {method} a {resource}
 @app.{convert_to_http_method(method)}("/v1/{resource}s/", response_model=List[schema.{resource_name_with_capital}])
@@ -92,9 +98,11 @@ async def {method}_{resource}({resource}: schema.{resource_name_with_capital}):
   {resource}s = []
   # do something
   return {resource}s
-""")
+"""
+        )
     elif method == "read":
-        write_to_output1(f"""
+        write_to_output1(
+            f"""
         
 # {method} a {resource}
 @app.{convert_to_http_method(method)}("/v1/{resource}s/{res}{resource}_id{source}", response_model=schema.{resource_name_with_capital})
@@ -102,9 +110,11 @@ async def {method}_{resource}({resource}_id : int):
   {resource} = []
   # do something
   return {resource}
-""")
+"""
+        )
     elif method == "update":
-        write_to_output1(f"""
+        write_to_output1(
+            f"""
         
 # {method} a {resource}
 @app.{convert_to_http_method(method)}("/v1/{resource}s/{res}{resource}_id{source}", response_model=List[schema.{resource_name_with_capital}])
@@ -112,9 +122,11 @@ async def {method}_{resource}({resource}_id : int):
   {resource}s = []
   # do something
   return {resource}s
-""")
+"""
+        )
     elif method == "delete":
-        write_to_output1(f"""
+        write_to_output1(
+            f"""
         
 # {method} a {resource}
 @app.{convert_to_http_method(method)}("/v1/{resource}s/{res}{resource}_id{source}", response_model=List[schema.{resource_name_with_capital}])
@@ -122,25 +134,30 @@ async def {method}_{resource}({resource}_id : int):
   {resource}s = []
   # do something
   return {resource}s
-""")
+"""
+        )
     else:
         raise ValueError("unrecognised method")
 
 
 def context_delimeter(context: str):
-    write_to_output1(f"""
+    write_to_output1(
+        f"""
 # =======================================#
 #      {context.lower().swapcase()}      #
 # =======================================#
-""")
+"""
+    )
 
 
 def context_delimeter2(context: str):
-    write_to_output2(f"""
+    write_to_output2(
+        f"""
 # =======================================#
 #      {context.lower().swapcase()}      #
 # =======================================#
-""")
+"""
+    )
 
 
 def create_db_schema(model: str):
@@ -148,7 +165,8 @@ def create_db_schema(model: str):
     for char in model[1:]:
         model_with_capital_letter += char
 
-    write_to_output2(f"""
+    write_to_output2(
+        f"""
 
 class {model_with_capital_letter}Base(BaseModel):
     title: str
@@ -166,25 +184,32 @@ class {model_with_capital_letter}({model_with_capital_letter}Base):
     class Config:
         orm_mode = True
 
-""")
+"""
+    )
+
 
 resources = []
 methods = ["create", "read", "update", "delete"]
 with open("models.py") as models:
-    content  = models.readlines()
+    content = models.readlines()
 
 for line in content:
     if line.find("class ") != -1:
-        resources.append(line.split("class ")[1].split("(Base)")[0].replace(" ","").lower())
+        resources.append(line.split("class ")[1].split("(Base)")[0].replace(" ", "").lower())
 
 database_schema = resources
 server_methods = []
-for method in methods:
-    for resource in resources:
+for resource in resources:
+    for method in methods:
         server_methods.append((method, resource))
 
-os.remove("app.py")
-os.remove("schema.py")
+print(server_methods)
+
+if Path("app.py").exists():
+    os.remove("app.py")
+
+if Path("schema.py").exists():
+    os.remove("schema.py")
 
 create_server_setup()
 resource_flag = ""
